fastapi/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
email-validator==2.0.0
pandas==2.0.1
python-dotenv==1.0.0
starlette==0.26.1